requirements.txt
click
requests
requests-futures
datapackage
